+++ linbox/tests/test-smith-form-adaptive.C
@@ -68,11 +68,11 @@ bool testRandom(const Ring& R,
 	VectorWrapper::ensureDim (x, stream1.n ());
 
 
-	int n = d. size();
+	int n = (int)d. size();
 
 	while (stream1) {
 
-		commentator.startIteration (stream1.j ());
+		commentator.startIteration ((unsigned)stream1.j ());
 
 		std::ostream &report = commentator.report (Commentator::LEVEL_IMPORTANT, INTERNAL_DESCRIPTION);
